pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = ""
authors = [
    {name = "takelele",email = "work.mnikulin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "geoalchemy2 (>=0.17.1,<0.18.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
]

[tool.poetry.dependencies]
python = "^3.10"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
